refactor: 헤더 및 푸터 폴더 구조 변경
