+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 
-import Footer from '@/components/Footer';
-import Header from '@/components/Header';
+import Footer from '@/pages/Footer';
+import Header from '@/pages/Header';
 
 interface AppProps {
   themeKey?: string;
